Move smartparens on new keybinding prefix "C-c s"
I don't use Smartparens, or better to say, I don't know how to use
Smartparens. But I have a lot of problem with Smartparens conflicting
with keybindings of other modes.
That is why I have removed all Smartparens related keybindings on
prefix key "C-c s", to keep it and have a chance to learn it and then
decide if I want to keep it or delete it.
When I learn how to use it, and does it have any benefit to my
workflow, then I'll reconsider if those keybindings are actually good
ones.