Remove ctrlf package
CTRLF package doesn't produces error when it gets to the last search.
This is makes it inconvenient for macros.
And since the reason why I went from Ivy to Selectrum, is to be more
close to the bare metal, I feel that this ctrlf package adds
unneccessary overlay.

I did like, however, ctrlf keybindings to go the the first seach M-<
and to the last search M->, and C-v and M-v respectively.